Inline MediaTypeCommon
This is an artifact of our multiplatform exploration.
