+++ okhttp/src/commonJvmAndroid/kotlin/okhttp3/MediaType.kt
@@ -16,12 +16,6 @@
 package okhttp3
 
 import java.nio.charset.Charset
-import okhttp3.internal.commonEquals
-import okhttp3.internal.commonHashCode
-import okhttp3.internal.commonParameter
-import okhttp3.internal.commonToMediaType
-import okhttp3.internal.commonToMediaTypeOrNull
-import okhttp3.internal.commonToString
 
 /**
  * An [RFC 2045][rfc_2045] Media Type, appropriate to describe the content type of an HTTP request
@@ -40,7 +34,7 @@ class MediaType internal constructor(
    */
   @get:JvmName("subtype") val subtype: String,
   /** Alternating parameter names with their values, like `["charset", "utf-8"]`. */
-  internal val parameterNamesAndValues: Array<String>,
+  private val parameterNamesAndValues: Array<String>,
 ) {
   /**
    * Returns the charset of this media type, or [defaultValue] if either this media type doesn't
@@ -60,7 +54,14 @@ class MediaType internal constructor(
    * Returns the parameter [name] of this media type, or null if this media type does not define
    * such a parameter.
    */
-  fun parameter(name: String): String? = commonParameter(name)
+  fun parameter(name: String): String? {
+    for (i in parameterNamesAndValues.indices step 2) {
+      if (parameterNamesAndValues[i].equals(name, ignoreCase = true)) {
+        return parameterNamesAndValues[i + 1]
+      }
+    }
+    return null
+  }
 
   @JvmName("-deprecated_type")
   @Deprecated(
@@ -82,13 +83,18 @@ class MediaType internal constructor(
    * Returns the encoded media type, like "text/plain; charset=utf-8", appropriate for use in a
    * Content-Type header.
    */
-  override fun toString(): String = commonToString()
+  override fun toString(): String = mediaType
 
-  override fun equals(other: Any?): Boolean = commonEquals(other)
+  override fun equals(other: Any?): Boolean = other is MediaType && other.mediaType == mediaType
 
-  override fun hashCode(): Int = commonHashCode()
+  override fun hashCode(): Int = mediaType.hashCode()
 
   companion object {
+    private const val TOKEN = "([a-zA-Z0-9-!#$%&'*+.^_`{|}~]+)"
+    private const val QUOTED = "\"([^\"]*)\""
+    private val TYPE_SUBTYPE = Regex("$TOKEN/$TOKEN")
+    private val PARAMETER = Regex(";\\s*(?:$TOKEN=(?:$TOKEN|$QUOTED))?")
+
     /**
      * Returns a media type for this string.
      *
@@ -96,12 +102,57 @@ class MediaType internal constructor(
      */
     @JvmStatic
     @JvmName("get")
-    fun String.toMediaType(): MediaType = commonToMediaType()
+    fun String.toMediaType(): MediaType {
+      val typeSubtype =
+        TYPE_SUBTYPE.matchAt(this, 0)
+          ?: throw IllegalArgumentException("No subtype found for: \"$this\"")
+      val type = typeSubtype.groupValues[1].lowercase()
+      val subtype = typeSubtype.groupValues[2].lowercase()
+
+      val parameterNamesAndValues = mutableListOf<String>()
+      var s = typeSubtype.range.last + 1
+      while (s < length) {
+        val parameter = PARAMETER.matchAt(this, s)
+        require(parameter != null) {
+          "Parameter is not formatted correctly: \"${substring(s)}\" for: \"$this\""
+        }
+
+        val name = parameter.groups[1]?.value
+        if (name == null) {
+          s = parameter.range.last + 1
+          continue
+        }
+
+        val token = parameter.groups[2]?.value
+        val value =
+          when {
+            token == null -> {
+              // Value is "double-quoted". That's valid and our regex group already strips the quotes.
+              parameter.groups[3]!!.value
+            }
+            token.startsWith('\'') && token.endsWith('\'') && token.length > 2 -> {
+              // If the token is 'single-quoted' it's invalid! But we're lenient and strip the quotes.
+              token.substring(1, token.length - 1)
+            }
+            else -> token
+          }
+
+        parameterNamesAndValues += name
+        parameterNamesAndValues += value
+        s = parameter.range.last + 1
+      }
+
+      return MediaType(this, type, subtype, parameterNamesAndValues.toTypedArray())
+    }
 
     /** Returns a media type for this, or null if this is not a well-formed media type. */
     @JvmStatic
     @JvmName("parse")
-    fun String.toMediaTypeOrNull(): MediaType? = commonToMediaTypeOrNull()
+    fun String.toMediaTypeOrNull(): MediaType? = try {
+      toMediaType()
+    } catch (_: IllegalArgumentException) {
+      null
+    }
 
     @JvmName("-deprecated_get")
     @Deprecated(
